changed most of the DB connection calls
I am unsure if the code will even run(sorry), I was honestly just modifying it in a python file in VScode that's why I copied over the whole file. 
need to implement/add to the db to account for all the functionality in the main file, which I outlined in the code itself:
1. need to implement the collaborator functionality in the db 
2. the code deals with hashed passwords which the db does not have the functionality for
3. functionality with blacklist, etc. shown in code is not reflect in db
I'm sure I am missing some things, but this is what I remember being issues. you can go through the code and look at my comments (labeled CB), I got up to the complaints section (around line 800-900). we should have a group meeting to discuss the issues addressed here ASAP and get to fixing the issues.